==> Notebooks/1_DataExtraction.ipynb <==
# Packages importation

from loky import ProcessPoolExecutor
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
from datetime import datetime
import re
from itertools import chain

# Functions definition

- This function is used to collect the `id_titles` of the content of the Netflix's database from the website `flixable.com`

def mapper_id_titles(url):
    response = requests.get(url)
    BS = (BeautifulSoup(response.text)
          .select("html body main div#filterContainer.container.pt-2 div.row div.col-sm-6.col-lg-3.mb-3 div.poster-container a.poster-link"))
    return [i.get('href') for i in BS]

# mapper_id_titles('https://fr.flixable.com/genre/{}/?min-rating=0&min-year=1920&max-year=2019&order=title&page=1')

- This function is used to collect data from the website `flixable.com`

def mapper_flixable(id_title):
    flixable = "https://fr.flixable.com" + id_title
    response = requests.get(flixable)
    
    nom = (BeautifulSoup(response.text)
           .select("html body main div.container.mt-4 div.row.mb-3 div.col-12 h1.mb-3")[0]
           .text.split("\n")[1].strip())
    
    inter = (BeautifulSoup(response.text)
             .select("div[class='col-md-8 mb-3'] span[class=imdbRatingPlugin] a"))
    if len(inter):
        lien_IMDb = inter[0]['href']
    else:
        lien_IMDb = ""
        
    inter = (BeautifulSoup(response.text)
             .select("html body main div.container.mt-4 div.row.mb-3 div.col-md-8.mb-3 div.mb-2 span a"))
    inter = ["".join(re.findall(r'.*/actor/.*',str(i))) for i in inter]
    inter = [i for i in inter if i!=""]
    casting = "|#|".join(["".join(re.findall(r'.*>(.*)</a>',i))for i in inter])
    
    date_ajout = (BeautifulSoup(response.text)
                  .select("div.mb-4 > span:nth-child(2)"))[0].text
    
    return (nom, lien_IMDb, casting, date_ajout)

# mapper_flixable("/title/70301585/")

- This function is used to collect data from the website `netflix.com/fr`

def mapper_netflix(id_title):
    netflix = "https://www.netflix.com/fr" + id_title
    response = requests.get(netflix)
    
    inter = (BeautifulSoup(response.text)
             .select("a.title-info-metadata-item"))
    genre1 = [i.text.strip().strip(',').replace('\xa0', ' ') for i in inter]
    
    inter = (BeautifulSoup(response.text)
             .select(".cell-genres > div:nth-child(2) > span"))
    genre2 = [i.text.strip().strip(',').replace('\xa0', ' ') for i in inter]
    
    genre = "|#|".join(set(genre1+genre2))
    
#     test = BeautifulSoup(response.text).select("#section-more-details")[0].text
#     if not re.search(r'.*Ce (?:programme|film) est\.\.\..*', test):
#         tag = ""
#     else:
#         inter = BeautifulSoup(response.text).select("div.more-details-cell div span[class$=tag]")
#         tag = "|#|".join([i.text.strip().strip(',').replace('\xa0', ' ') for i in inter])
        
#     inter = BeautifulSoup(response.text).select(".link-container a")
#     relative = "|#|".join([i['href'].strip('/fr') for i in inter])
    
    return genre#(, relative, tag)

# mapper_netflix("/title/80997613/")

- This function is used to collect data from the website `imdb.com`

def mapper_rating_IMDb(id_IMDb):
    if id_IMDb=="":
        return ""
    else:
        response = requests.get(id_IMDb)
        BS = (BeautifulSoup(response.text)
              .select(".ratingValue > strong:nth-child(1) > span:nth-child(1)"))
        if len(BS):
            return float(BS[0].text)
        else:
            return ""

# mapper_rating_IMDb("https://www.imdb.com/title/tt3656200/?ref_=plg_rt_1")

- These functions are used to parallelize the functions above

def all_id_title(key):
    with ProcessPoolExecutor() as epool:
        mapped_values = epool.map(mapper_id_titles, key)
    return list(chain(*mapped_values))

def all_flixable(key):
    with ProcessPoolExecutor() as epool:
        mapped_values_flixable = epool.map(mapper_flixable, key)
    return list(mapped_values_flixable)

def all_netflix(key):
    with ProcessPoolExecutor() as epool:
        mapped_values_netflix = epool.map(mapper_netflix, key)
    return list(mapped_values_netflix)

def all_rating_IMDb(key):
    with ProcessPoolExecutor() as epool:
        mapped_values_rating_IMDb = epool.map(mapper_rating_IMDb, key)
    return list(mapped_values_rating_IMDb)

# Data extraction

**Collecting the amount of pages on the website `flixable.com`**

url = "https://fr.flixable.com/genre/{}/?min-rating=0&min-year=1920&max-year=2020&order=title#filterContainer"
# Here, it's the Netflix's database of France
# For the Netflix's database of USA : https://www.flixable.com/
# For the Netflix's database of Germany : https://de.flixable.com/

types = ["films", "series-tv"]
number_of_pages = {}

for elem in types:
    response = requests.get(url.format(elem))
    BS = BeautifulSoup(response.text).select("li.page-item:nth-child(7) > a:nth-child(1)")[0]
    number_of_pages[elem] = int(BS.get_text())

number_of_pages

**Collecting all URLs from the `flixable.com` website**

url = "https://fr.flixable.com/genre/{}/?min-rating=0&min-year=1920&max-year=2019&order=title&page="
urls = {}

for elem in types:
    N = number_of_pages[elem]
    urls[elem] = [url.format(elem) + str(i+1) for i in range(N)]

for key, value in urls.items():
    print("{} : {}\n".format(key,value[:2]))

**Collecting all `id_titles`**

%%time
id_titles = {}
for key in types:
    id_titles[key] = all_id_title(urls[key])

for key, value in id_titles.items():
    print("For the {}, the amount is {}\n".format(key,len(value)))

for key, value in id_titles.items():
    print("{} : {}\n".format(key,value[:2]))

urls_films = id_titles['films']
urls_series = id_titles['series-tv']
numero_titres = urls_films + urls_series

**Collecting data from `flixable.com`**

%%time
data_flixable = {}
for key in types:
    data_flixable[key] = all_flixable(id_titles[key])

for key, value in data_flixable.items():
    print("{} : {}\n".format(key,value[:2]))

infos_flixable = []

for elem in data_flixable.values():
    infos_flixable += elem

title = []
urls_IMDb = []
casting = []
date_ajout = []

for elem in infos_flixable:
    title.append(elem[0])
    urls_IMDb.append(elem[1])
    casting.append(elem[2])
    date_ajout.append(elem[3])

types_media = ['film']*len(urls_films) + ['série']*len(urls_series)
types_media[:5],types_media[-5:]

**Collecting data from `netflix.com/fr`**

Doesn't work because there are too many requests at the same time. I suppose Netflix blocks these requests.

```python
%%time
data_netflix = {}
for key in types:
    data_netflix[key] = all_netflix(id_titles[key])
```

%%time
infos_netflix = []
URLS = urls_films + urls_series

for elem in URLS:
    infos_netflix.append(mapper_netflix(elem))

infos_netflix[:5]

# genres = []
# tags = []
# relative = []
genres = [i for i in infos_netflix]
# for elem in infos_netflix:
#     genres.append(elem[0])
#     tags.append(elem[1])
#     relative.append(elem[2])

**Collecting data from `imdb.com`**

%%time
rating = all_rating_IMDb(urls_IMDb)
rating[:10]

# Dataframe creation

# data = {"Titre": title, "Genres Netflix" : genres,
#         "Tags Netflix" : tags, "Casting" : casting,
#         "Note IMDb" : rating, "Films/Séries liés" : relative,
#         "Date d'ajout" : date_ajout, "Lien Netflix" : numero_titres,
#         "Type" : types_media}

data = {"Titre": title, "Genres Netflix" : genres,
        "Casting" : casting, "Note IMDb" : rating,
        "Date d'ajout" : date_ajout, "Lien Netflix" : numero_titres,
        "Type" : types_media}

BDD_Netflix = pd.DataFrame(data)
BDD_Netflix = BDD_Netflix.loc[BDD_Netflix['Genres Netflix']!='',:]

def string_to_datetime(valeur):
    return datetime.strptime(valeur, '%d/%m/%Y')

BDD_Netflix["Date d'ajout"] = BDD_Netflix["Date d'ajout"].apply(string_to_datetime)       

def list_to_string(valeur):
    if valeur=="":
        return np.nan
    else:
        return valeur

#BDD_Netflix["Tags Netflix"] = BDD_Netflix["Tags Netflix"].apply(list_to_string)
BDD_Netflix["Note IMDb"] = BDD_Netflix["Note IMDb"].apply(list_to_string)
BDD_Netflix["Casting"] = BDD_Netflix["Casting"].apply(list_to_string)
#BDD_Netflix["Films/Séries liés"] = BDD_Netflix["Films/Séries liés"].apply(list_to_string)

BDD_Netflix.head()

BDD_Netflix.tail()

BDD_Netflix.shape

# Exportation to `csv` and `pickle`

BDD_Netflix.to_csv('../Data/BDD_Netflix.csv', sep = ';', index = False, encoding = 'utf-8', na_rep = 'NA')

BDD_Netflix.to_pickle('../Data/BDD_Netflix.pkl')

# Sources

- https://fr.flixable.com


- https://www.netflix.com/fr/


- https://pandas.pydata.org/pandas-docs/stable/reference/frame.html


- https://www.crummy.com/software/BeautifulSoup/bs4/doc/


- https://docs.python.org/2/library/re.html
