==> Notebooks/2_LinksCreation.ipynb <==
# Packages importation

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from collections import Counter
import pyspark
from pyspark.sql import SQLContext

# Functions definition

def StringToList(x):
    if isinstance(x, float):
        return x
    else:
        return x.split("|#|")
    
def DeleteCaractere(x):
    if isinstance(x, float):
        return x
    else:
        return x.replace('"',"'")

# This function add +1 to the value of a dictionary's key
def appendOrAdd_dico(dico, key):
    if key in dico:
        dico[key]+=1
    else:
        dico[key]=1
    return dico

# First, this function filter and keep all id_titles (`Lien Netflix`) where the parameter `mot` appears in the columns `col_name`.
# Then, it creates links between all of these id_titles and add them to a dictionary.
# Finally, it returns the dictionary of links between id_titles (ex : {'/title/70060018/	/title/70270776/':2})
def create_links(mot, col_name, dico, BDD = "BDD_NetflixSQL"):
    request = f"""SELECT `Lien Netflix`
                  FROM {BDD}
                  WHERE (`{col_name}` LIKE "{mot}")
                     OR (`{col_name}` LIKE "%|{mot}")
                     OR (`{col_name}` LIKE "%|{mot}|%")
                     OR (`{col_name}` LIKE "{mot}|%")"""
    
    inter = sqlContext.sql(request).toPandas()['Lien Netflix']
    
    for i in range(len(inter)):
        for j in range(i+1,len(inter)):
            name1 = inter[i]
            name2 = inter[j]
            key = f"{name1}\t{name2}"
            dico = appendOrAdd_dico(dico, key)
            
    return dico

# Data importation

- When working with Apache Spark we invoke methods on an object which is an instance of the `pyspark.SparkContext` context.

- Typically, an instance of this object will be created automatically for you and assigned to the variable `sc`.

sc = pyspark.SparkContext(master="local[*]", appName="FirstExample")
sqlContext = SQLContext(sc)

BDD_Netflix_Spark = sqlContext.read.csv("../Data/BDD_Netflix.csv", sep = ";", header = True, inferSchema=True)
BDD_Netflix_Spark.createOrReplaceTempView("BDD_NetflixSQL")

BDDFilm = BDD_Netflix_Spark.filter(BDD_Netflix_Spark.Type=="film")
BDDFilm.createOrReplaceTempView("BDD_Film")

BDDSerie = BDD_Netflix_Spark.filter(BDD_Netflix_Spark.Type=="série")
BDDSerie.createOrReplaceTempView("BDD_Serie")

BDD_Netflix = pd.read_csv('../Data/BDD_Netflix.csv', sep = ';')
BDD_Netflix = BDD_Netflix.set_index('Lien Netflix')
BDD_Netflix['Casting'] = BDD_Netflix['Casting'].apply(DeleteCaractere)

BDD_Film = BDD_Netflix.loc[BDD_Netflix['Type']=='film',:]
BDD_Serie = BDD_Netflix.loc[BDD_Netflix['Type']=='série',:]

BDD_Netflix.head()

```python
BDD_Netflix['Genres Netflix'] = BDD_Netflix['Genres Netflix'].apply(StringToList)
BDD_Netflix['Tags Netflix'] = BDD_Netflix['Tags Netflix'].apply(StringToList)
BDD_Netflix['Casting'] = BDD_Netflix['Casting'].apply(StringToList)
BDD_Netflix['Films/Séries liés'] = BDD_Netflix['Films/Séries liés'].apply(StringToList)

BDD_Netflix.head()
```

**Number of appearances of the most common "genres/tags/actrices-acteurs/contenu"**

Liste_Genres = "|#|".join(BDD_Netflix['Genres Netflix']).split("|#|")
Liste_Genres_film = "|#|".join(BDD_Film['Genres Netflix']).split("|#|")
Liste_Genres_serie = "|#|".join(BDD_Serie['Genres Netflix']).split("|#|")

#Liste_Tags = "|#|".join(BDD_Netflix['Tags Netflix'].dropna()).split("|#|")

Liste_Casting = "|#|".join(BDD_Netflix['Casting'].dropna()).split("|#|")
Liste_Casting_film = "|#|".join(BDD_Film['Casting'].dropna()).split("|#|")
Liste_Casting_serie = "|#|".join(BDD_Serie['Casting'].dropna()).split("|#|")

#Liste_Liaisons = "|#|".join(BDD_Netflix['Films/Séries liés'].dropna()).split("|#|")

dico_genres = Counter(Liste_Genres)
dico_genres_film = Counter(Liste_Genres_film)
dico_genres_serie = Counter(Liste_Genres_serie)

#dico_tags = Counter(Liste_Tags)

dico_casting = Counter(Liste_Casting)
dico_casting_film = Counter(Liste_Casting_film)
dico_casting_serie = Counter(Liste_Casting_serie)

#dico_liaisons = Counter(Liste_Liaisons)

We remove the three most common genres to avoid having a hyperconnected graph

genres_3_most_common = dico_genres.most_common(3)
genres_3_most_common = [i[0] for i in genres_3_most_common]

genres_film_3_most_common = dico_genres_film.most_common(3)
genres_film_3_most_common = [i[0] for i in genres_film_3_most_common]

genres_serie_3_most_common = dico_genres_serie.most_common(3)
genres_serie_3_most_common = [i[0] for i in genres_serie_3_most_common]

```python
>>>dico_genres.most_common(10)
[('Comédies', 958),
 ('Drames', 593),
 ('Films documentaires', 438),
 ('Séries US', 389),
 ('Action et aventure', 369),
 ('Séries comiques', 352),
 ('Pour enfants', 329),
 ('Films inspirés de livres', 323),
 ('Films romantiques', 309),
 ('Séries dramatiques', 266)]
```

***\"Factors\"*** **de genres/tags/actrices-acteurs/contenu**

Liste_Genres = np.unique(Liste_Genres)
# If you want to delete the 3 most common genres, use the following line
#Liste_Genres = [i for i in Liste_Genres if i not in genres_3_most_common]
Liste_Genres_film = np.unique(Liste_Genres_film)
#Liste_Genres_film = [i for i in Liste_Genres_film if i not in genres_film_3_most_common]
Liste_Genres_serie = np.unique(Liste_Genres_serie)
#Liste_Genres_serie = [i for i in Liste_Genres_serie if i not in genres_serie_3_most_common]

#Liste_Tags = np.unique(Liste_Tags)

Liste_Casting = np.unique(Liste_Casting)
Liste_Casting_film = np.unique(Liste_Casting_film)
Liste_Casting_serie = np.unique(Liste_Casting_serie)

#Liste_Liaisons = np.unique(Liste_Liaisons)

## Links creation between contents according to genres

These chunks create 3 dictionaries (one for the full database, one for the movie database, one for the serie database). Keys of these dictionaries are the link between 2 contents and the values are the numbers of genres they have in common.

%%time
col_name = "Genres Netflix"
dico_links_genres = {}
for genre in Liste_Genres:
    dico_links_genres = create_links(genre, col_name, dico_links_genres)
    #print(genre)

sorted(dico_links_genres.items(), key=lambda x: -x[1])[:10]

len(dico_links_genres)

%%time
col_name = "Genres Netflix"
BDD = "BDD_Film"
dico_links_genres_film = {}
for genre in Liste_Genres_film:
    dico_links_genres_film = create_links(genre, col_name, dico_links_genres_film, BDD)
    #print(genre)

sorted(dico_links_genres_film.items(), key=lambda x: -x[1])[:10]

len(dico_links_genres_film)

%%time
col_name = "Genres Netflix"
BDD = "BDD_Serie"
dico_links_genres_serie = {}
for genre in Liste_Genres_serie:
    dico_links_genres_serie = create_links(genre, col_name, dico_links_genres_serie, BDD)
    #print(genre)

sorted(dico_links_genres_serie.items(), key=lambda x: -x[1])[:10]

len(dico_links_genres_serie)

## Links creation between contents according to tags

```python
%%time
col_name = "Tags Netflix"
dico_links_tags = {}
for tag in Liste_Tags:
    dico_links_tags = create_links(tag, col_name, dico_links_tags)
    #print(tag)
```

```python
sorted(dico_links_tags.items(), key=lambda x: -x[1])[:10]
```

## Links creation between contents according to the casting

Choice to make on the number of appearances of the actor or actress (for example : if the actor/actress appears only once in a Netflix content, we remove him/her from the `Liste_Casting` in order to not launch the function for nothing. Here, we decided to keep only actors/actresses that appears more than 5 times in the Netflix database (arbitrary choice).

i = 0
Liste_Casting_sup1 = []
dico_casting = sorted(dico_casting.items(), key=lambda x: -x[1])

while dico_casting[i][1]>5:
    Liste_Casting_sup1.append(dico_casting[i][0])
    i=i+1

i = 0
Liste_Casting_sup1_film = []
dico_casting_film = sorted(dico_casting_film.items(), key=lambda x: -x[1])

while dico_casting_film[i][1]>5:
    Liste_Casting_sup1_film.append(dico_casting_film[i][0])
    i=i+1
    
i = 0
Liste_Casting_sup1_serie = []
dico_casting_serie = sorted(dico_casting_serie.items(), key=lambda x: -x[1])

while dico_casting_serie[i][1]>5:
    Liste_Casting_sup1_serie.append(dico_casting_serie[i][0])
    i=i+1

len(Liste_Casting_sup1),len(Liste_Casting_sup1_film),len(Liste_Casting_sup1_serie)

These chunks create 3 dictionaries (one for the full database, one for the movie database, one for the serie database). Keys of these dictionaries are the link between 2 contents and the values are the numbers of actors/actresses they have in common.

%%time
col_name = "Casting"
dico_links_casting = {}
for casting in Liste_Casting_sup1:
    dico_links_casting = create_links(casting, col_name, dico_links_casting)
    #print(casting)

sorted(dico_links_casting.items(), key=lambda x: -x[1])[:10]

len(dico_links_casting)

%%time
col_name = "Casting"
BDD = "BDD_Film"
dico_links_casting_film = {}
for casting in Liste_Casting_sup1_film:
    dico_links_casting_film = create_links(casting, col_name, dico_links_casting_film,BDD)
    #print(casting)

sorted(dico_links_casting_film.items(), key=lambda x: -x[1])[:10]

len(dico_links_casting_film)

%%time
col_name = "Casting"
BDD = "BDD_Serie"
dico_links_casting_serie = {}
for casting in Liste_Casting_sup1_serie:
    dico_links_casting_serie = create_links(casting, col_name, dico_links_casting_serie,BDD)
    #print(casting)

sorted(dico_links_casting_serie.items(), key=lambda x: -x[1])[:10]

len(dico_links_casting_serie)

## Links creation between contents according to the column `Films/Séries liés`

```python
%%time
col_name = "Films/Séries liés"
dico_links_liaison = {}
for liaison in Liste_Liaisons:
    dico_links_liaison = create_links(liaison, col_name, dico_links_liaison)
    #print(liaison)
```

```python
sorted(dico_links_liaison.items(), key=lambda x: -x[1])[:10]
```

## Gatherings of different dictionaries

%%time
# We have decided to only take into account the links according to genres and the casting
dicoAll = [dico_links_genres, dico_links_casting]
dicoAll_film = [dico_links_genres_film, dico_links_casting_film]
dicoAll_serie = [dico_links_genres_serie, dico_links_casting_serie]

# sum the values with same keys (a weight could be taken into consideration but not here)
counter = Counter() 
for d in dicoAll:  
    counter.update(d) 
      
result = dict(counter)

counter_film = Counter() 
for d in dicoAll_film:  
    counter_film.update(d) 
      
result_film = dict(counter_film) 

counter_serie = Counter() 
for d in dicoAll_serie:  
    counter_serie.update(d) 
      
result_serie = dict(counter_serie) 

sorted(result.items(), key=lambda x: -x[1])[:10]

# Nodes weight

Here, the weight of a node is it rating on IMDb

dico_nodes = {link:BDD_Netflix["Note IMDb"][link] for link in BDD_Netflix.index}
dico_nodes_film = {link:BDD_Film["Note IMDb"][link] for link in BDD_Film.index}
dico_nodes_serie = {link:BDD_Serie["Note IMDb"][link] for link in BDD_Serie.index}

dico_nodes

# Relation between id_title et name of the content

dico_title = {link:BDD_Netflix["Titre"][link] for link in BDD_Netflix.index}
dico_title_film = {link:BDD_Film["Titre"][link] for link in BDD_Film.index}
dico_title_serie = {link:BDD_Serie["Titre"][link] for link in BDD_Serie.index}

dico_title

# Exportation

- The links

with open("../Data/Liaisons.txt","w", encoding = "utf-8") as f:
    for key, value in result.items():
        f.write(f"{key}\t{value}\n")
        
with open("../Data/Liaisons_film.txt","w", encoding = "utf-8") as f:
    for key, value in result_film.items():
        f.write(f"{key}\t{value}\n")
        
with open("../Data/Liaisons_serie.txt","w", encoding = "utf-8") as f:
    for key, value in result_serie.items():
        f.write(f"{key}\t{value}\n")

- The nodes weight

with open("../Data/Nodes_weight.txt","w", encoding = "utf-8") as f:
    for key, value in dico_nodes.items():
        f.write(f"{key}\t{value}\n")
        
with open("../Data/Nodes_weight_film.txt","w", encoding = "utf-8") as f:
    for key, value in dico_nodes_film.items():
        f.write(f"{key}\t{value}\n")
        
with open("../Data/Nodes_weight_serie.txt","w", encoding = "utf-8") as f:
    for key, value in dico_nodes_serie.items():
        f.write(f"{key}\t{value}\n")

- The titles

with open("../Data/Titles.txt","w", encoding = "utf-8") as f:
    for key, value in dico_title.items():
        f.write(f"{key}\t{value}\n")

with open("../Data/Titles_film.txt","w", encoding = "utf-8") as f:
    for key, value in dico_title_film.items():
        f.write(f"{key}\t{value}\n")
        
with open("../Data/Titles_serie.txt","w", encoding = "utf-8") as f:
    for key, value in dico_title_serie.items():
        f.write(f"{key}\t{value}\n")
